refactor to page and to standard page (#120)
