+++ components/Page/asStandardPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import PlayerWrapper from "./FooterPlayer/PlayerWrapper/PlayerWrapper";
-import { Navbar } from "./ui/Page";
+import { Navbar } from "../ui/Page";
+import styles from "./Page.module.scss";
 
 export const asStandardPage = (Page: React.ElementType) => {
   // @ts-expect-error
@@ -8,7 +8,7 @@ export const asStandardPage = (Page: React.ElementType) => {
     return (
       <>
         <Navbar />
-        <div style={{ paddingBottom: 15 }}>{page}</div>
+        <div className={styles.standardPage}>{page}</div>
       </>
     );
   };
