+++ styles/globals.scss
@@ -45,6 +45,8 @@ a {
 	--shadow: 3px 3px 4px #000000;
 	--inverted-shadow: -3px -3px 4px #000000;
 
+	--player-height: 110px;
+
 	--navbar-width: 200px;
 
 	--SPACE1: 15px;
